backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
pandas
numpy
numpy_financial


